Rust: generate the extractor
