+++ rust/generate-schema/src/main.rs
@@ -402,14 +402,32 @@ fn get_fields(node: &AstNodeSrc) -> Vec<FieldInfo> {
 
 fn write_extractor(grammar: &AstSrc) -> std::io::Result<String> {
     let mut buf: Vec<u8> = Vec::new();
+    writeln!(
+        buf,
+        "use crate::generated;
+use super::base::{{TextValue, Translator}};
+use crate::trap::{{Label, TrapId}};
+use ra_ap_syntax::ast;
+use ra_ap_syntax::ast::{{
+    HasArgList, HasAttrs, HasGenericArgs, HasGenericParams, HasLoopBody, HasModuleItem, HasName,
+    HasTypeBounds, HasVisibility, RangeItem,
+}};
 
+impl Translator {{
+    fn emit_else_branch(&mut self, node: ast::ElseBranch) -> Label<generated::Expr> {{
+        match node {{
+            ast::ElseBranch::IfExpr(inner) => self.emit_if_expr(inner).into(),
+            ast::ElseBranch::Block(inner) => self.emit_block_expr(inner).into(),
+        }}
+    }}\n"
+    )?;
     for node in &grammar.enums {
         let type_name = &node.name;
         let class_name = class_name(&node.name);
 
         writeln!(
             buf,
-            "    fn emit_{}(&mut self, node: ast::{}) -> Label<generated::{}> {{",
+            "    pub(crate) fn emit_{}(&mut self, node: ast::{}) -> Label<generated::{}> {{",
             to_lower_snake_case(type_name),
             type_name,
             class_name
@@ -434,7 +452,7 @@ fn write_extractor(grammar: &AstSrc) -> std::io::Result<String> {
 
         writeln!(
             buf,
-            "    fn emit_{}(&mut self, node: ast::{}) -> Label<generated::{}> {{",
+            "    pub(crate) fn emit_{}(&mut self, node: ast::{}) -> Label<generated::{}> {{",
             to_lower_snake_case(type_name),
             type_name,
             class_name
@@ -491,6 +509,7 @@ fn write_extractor(grammar: &AstSrc) -> std::io::Result<String> {
 
         writeln!(buf, "    }}\n")?;
     }
+    writeln!(buf, "}}")?;
     Ok(String::from_utf8_lossy(&buf).into_owned())
 }
 
@@ -523,13 +542,21 @@ fn main() -> std::io::Result<()> {
     });
     let schema = write_schema(&grammar, super_types)?;
     let schema_path = PathBuf::from("../schema/ast.py");
-    let extractor = write_extractor(&grammar)?;
-    print!("{}", extractor);
     codegen::ensure_file_contents(
         crate::flags::CodegenType::Grammar,
         &schema_path,
         &schema,
         false,
     );
+
+    let extractor = write_extractor(&grammar)?;
+    let extractor_path = PathBuf::from("../extractor/src/translate/generated.rs");
+    codegen::ensure_file_contents(
+        crate::flags::CodegenType::Grammar,
+        &extractor_path,
+        &extractor,
+        false,
+    );
+
     Ok(())
 }
